server/requirements.txt
flask
flask-cors
pillow
torch
torchvision
python-dotenv
gunicorn
numpy